services/email_mfa/requirements.txt
playwright==1.40
pyyaml==6.0.3
fastapi==0.125.0
pydantic==2.12.5
uvicorn==0.38.0
setuptools==44.1.1
pygments==2.19.2